core-refactor: rename usePathParam
