+++ packages/core/src/router/index.ts
@@ -4,5 +4,5 @@ export { default as useNavigate } from '@router/hook/useNavigate'
 export { default as useOutlet } from '@router/hook/useOutlet'
 export { default as usePathEffect } from '@router/hook/usePathEffect'
 export { default as usePathname } from '@router/hook/usePathname'
-export { default as usePathParam } from '@router/hook/usePathParam'
+export { default as usePathParams } from '@router/hook/usePathParams'
 export { default as useQueryParams } from '@router/hook/useQueryParams'
